backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-multipart
aiofiles
PyJWT
passlib[bcrypt]
